+++ src/main/g8/app/testkit/src/main/scala/myapp/utility/scalatest/SpecAssertions.scala
@@ -7,7 +7,6 @@ trait SpecAssertions {
   /** 条件が満たされているかチェックします。
     * 条件に違反している場合は [[java.lang.AssertionError]] をスローします。
     */
-  val expect = new Expecty {
-    override val failEarly: Boolean = false
-  }
+  val expect = new Expecty()
+
 }