+++ src/main/g8/app/testkit/src/test/scala/myapp/utility/scalatest/StandardSpecExample.scala
@@ -16,25 +16,6 @@ class StandardSpecExample extends StandardSpec {
         */
     }
 
-    "複数項目チェックする場合" in {
-      expect {
-        "1" === "1"
-        "2" === "2"
-      }
-
-      /** エラー出力の例：
-        * java.lang.AssertionError:
-        *
-        * "2" === "1"
-        * |   |
-        * 2   false
-        *
-        * "3" === "2"
-        * |   |
-        * 3   false
-        */
-    }
-
     "条件を満たさない場合はAssertionErrorになる" in {
       assertThrows[AssertionError] {
         expect("1" === "2")